encapsulate load data creation
