+++ packages/dds/test-dds-utils/src/index.ts
@@ -8,7 +8,6 @@ export {
 	AddClient,
 	BaseOperation,
 	ChangeConnectionState,
-	Client,
 	ClientSpec,
 	createDDSFuzzSuite,
 	DDSFuzzModel,
@@ -21,3 +20,4 @@ export {
 } from "./ddsFuzzHarness";
 export { createSnapshotSuite, ISnapshotSuite } from "./ddsSnapshotHarness";
 export { MinimizationTransform } from "./minification";
+export { Client } from "./clientLoading";
